Add longitudes check for ABI
**Description**
This PR adds a check for longitudes that are 0-360 and converts them to -180 to 180. This is to ensure correct coverage for satellites such as GOES-18.

**Issue closed**
None

**List of modified files**
```
src/goes_abi_mod.F90
test/abi/flist.txt_g18
test/abi/namelist.obs2mpas_g18
test/abi/plot.py
```
**Tests**
Tests were conducted on Derecho using GOES-18 data with superobbing on a 15 km MPAS mesh. The figures below show the coverage before and after the fix for the cloud amount variable on 2025071823. 
_Before_
<img width="2015" height="1210" alt="cloudAmount_g18_2025071823" src="https://github.com/user-attachments/assets/e54600de-a375-4b45-9c4e-67994107e01e" />
_After_
<img width="2015" height="1210" alt="cldamount_so_ahi" src="https://github.com/user-attachments/assets/9dc36048-4703-4b96-a734-f860a212cc87" />